fix: divide SlugSelectionComponents into smaller files
